requirements.txt
pylint>=1.4.3
gunicorn>=19.3.0
RPi.GPIO>=0.5.11
jedi>=0.9.0
django-cors-headers>=1.1.0
pillow>=2.8.2
